+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [13.20.2](https://kaos.sh/ek/13.20.2)
 
 - **`[knf/validators/time]`** Added timezone validator
+- **`[knf/validators]`** Code refactoring
 
 ### [13.20.1](https://kaos.sh/ek/13.20.1)